ref: refactor gitlab setup to multiple files
