+++ workspaces/backstage/e2e-tests/support/gitlab/scaffolder-test-setup.ts
@@ -4,31 +4,15 @@ import path from "node:path";
 
 import { requireEnv } from "@red-hat-developer-hub/e2e-test-utils/utils";
 
-import { GitLabApiHelper } from "./api/gitlab-api-helper.js";
+import { GitLabApiHelper } from "../api/gitlab-api-helper.js";
 
-export const GITLAB_EVENTS_CATALOG_TOKEN = "test-token";
 export const GITLAB_SCAFFOLDER_PARENT_GROUP = "rhdh-qe-test";
-
-const GITLAB_EVENTS_RHDH_CONFIG = {
-  auth: "keycloak" as const,
-  appConfig: "tests/config/gitlab-events/app-config-rhdh.yaml",
-  secrets: "tests/config/gitlab-events/rhdh-secrets.yaml",
-  dynamicPlugins: "tests/config/gitlab-events/dynamic-plugins.yaml",
-  valueFile: "tests/config/gitlab-events/value-file.yaml",
-};
 
 const GITLAB_SCAFFOLDER_RHDH_CONFIG = {
   auth: "guest" as const,
   appConfig: "tests/config/gitlab-scaffolder/app-config-rhdh.yaml",
   secrets: "tests/config/gitlab-scaffolder/rhdh-secrets.yaml",
   dynamicPlugins: "tests/config/gitlab-scaffolder/dynamic-plugins.yaml",
-};
-
-/** Worker fixture shape used by GitLab events E2E suites */
-export type GitLabEventsRhdhWorker = {
-  configure: (options: typeof GITLAB_EVENTS_RHDH_CONFIG) => Promise<void>;
-  deploy: () => Promise<void>;
-  rhdhUrl: string;
 };
 
 /** Worker fixture shape used by GitLab scaffolder E2E suite */
@@ -47,33 +31,8 @@ export interface GitLabScaffolderSharedState {
   publishCompleted: boolean;
 }
 
-export function requireGitLabEventsVaultEnv(): void {
-  requireEnv("VAULT_EVENTS_GITLAB_TOKEN");
-  requireEnv("VAULT_EVENTS_GITLAB_HOST");
-  requireEnv("VAULT_EVENTS_GITLAB_PARENT_ORG");
-  requireEnv("VAULT_GITLAB_WEBHOOK_SECRET");
-}
-
 export function requireGitLabDiscoveryVaultEnv(): void {
   requireEnv("VAULT_GITLAB_TOKEN_DECODED");
-}
-
-/**
- * Validates vault/GitLab env, initializes {@link GitLabApiHelper}, and returns a
- * unique resource prefix for this run.
- */
-export function bootstrapGitLabEventsApiClient(): string {
-  requireGitLabEventsVaultEnv();
-  const host = process.env.VAULT_EVENTS_GITLAB_HOST;
-  const token = process.env.VAULT_EVENTS_GITLAB_TOKEN;
-  if (typeof host !== "string" || host.length === 0) {
-    throw new TypeError("VAULT_EVENTS_GITLAB_HOST must be set");
-  }
-  if (typeof token !== "string" || token.length === 0) {
-    throw new TypeError("VAULT_EVENTS_GITLAB_TOKEN must be set");
-  }
-  GitLabApiHelper.init(`https://${host}`, token);
-  return GitLabApiHelper.generateTestPrefix();
 }
 
 /**
@@ -191,53 +150,9 @@ export function deleteGitLabScaffolderSharedState(
   }
 }
 
-export async function deployGitLabEventsHub(
-  rhdh: GitLabEventsRhdhWorker,
-): Promise<{ rhdhUrl: string; catalogToken: string }> {
-  await rhdh.configure(GITLAB_EVENTS_RHDH_CONFIG);
-  await rhdh.deploy();
-  return { rhdhUrl: rhdh.rhdhUrl, catalogToken: GITLAB_EVENTS_CATALOG_TOKEN };
-}
-
 export async function deployGitLabScaffolderHub(
   rhdh: GitLabScaffolderRhdhWorker,
 ): Promise<void> {
   await rhdh.configure(GITLAB_SCAFFOLDER_RHDH_CONFIG);
   await rhdh.deploy();
 }
-
-export async function prepareGitLabParentGroup(
-  parentGroupPath: string | undefined,
-): Promise<{
-  parentGroupId: number;
-  parentGroupPath: string;
-}> {
-  const parentGroup = await GitLabApiHelper.getGroupByPath(parentGroupPath);
-  await GitLabApiHelper.cleanupStaleResources(parentGroup.id, "e2e-", 1);
-  return {
-    parentGroupId: parentGroup.id,
-    parentGroupPath: parentGroup.full_path,
-  };
-}
-
-export async function prepareGitLabEventsParentGroup(): Promise<{
-  parentGroupId: number;
-  parentGroupPath: string;
-}> {
-  return prepareGitLabParentGroup(process.env.VAULT_EVENTS_GITLAB_PARENT_ORG);
-}
-
-export async function runGitLabCleanupSafely(
-  cleanup: () => Promise<void>,
-): Promise<void> {
-  try {
-    await cleanup();
-  } catch (error) {
-    console.warn(
-      `Cleanup error: ${error instanceof Error ? error.message : String(error)}`,
-    );
-  }
-}
-
-/** @deprecated Use {@link runGitLabCleanupSafely} */
-export const runGitLabEventsCleanupSafely = runGitLabCleanupSafely;
